blog feature code review
