+++ app/Models/Tag.php
@@ -4,6 +4,7 @@ namespace App\Models;
 
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
+use Illuminate\Database\Eloquent\Relations\BelongsToMany;
 
 class Tag extends Model
 {
@@ -11,8 +12,9 @@ class Tag extends Model
 
     protected $fillable = ['name'];
 
-    public function blogs()
+
+    public function blogs(): BelongsToMany
     {
-        return $this->belongsToMany(Blog::class, 'blog_tag');
+        return $this->belongsToMany(Blog::class);
     }
 }
